refactor: deduplicate collect_account_types logic between CLI and proc-macro paths
Move the shared account type scanning implementation into a new
spel_framework_core::account_types module (gated behind idl-gen feature).
Both the CLI path (spel generate-idl) in spel-framework-core and the
proc-macro path (lez_program, generate_idl!) now use this single source
of truth.

- spel-framework-core/src/account_types.rs: new shared module
- spel-framework-core/src/idl_gen.rs: removed ~240 lines of duplicated code,
  imports from account_types instead
- spel-framework-macros/src/account_types.rs: re-exports from core
- spel-framework-core/src/lib.rs: exports account_types behind idl-gen
