+++ spel-framework-core/src/lib.rs
@@ -10,6 +10,7 @@ pub mod pda;
 pub mod validation;
 
 #[cfg(feature = "idl-gen")]
+pub mod account_types;
 pub mod idl_gen;
 
 pub mod prelude {
